Species selector for pils
* bumped components and react components with species selector field type
* add new step after training modules for species
* updated species joining logic for top level pil
* show species on task reviewn
