+++ package.json
@@ -26,14 +26,14 @@
   },
   "homepage": "https://github.com/UKHomeOffice/asl-pages#readme",
   "dependencies": {
-    "@asl/components": "^4.3.3",
+    "@asl/components": "^4.4.1",
     "@asl/constants": "^0.3.2",
     "@asl/dictionary": "^1.1.4",
     "@asl/projects": "^5.12.6",
     "@asl/service": "^7.8.0",
     "@joefitter/docx": "^4.7.0",
     "@ukhomeoffice/frontend-toolkit": "^2.10.0",
-    "@ukhomeoffice/react-components": "^0.8.3",
+    "@ukhomeoffice/react-components": "^0.8.5",
     "body-parser": "^1.18.3",
     "classnames": "^2.2.5",
     "csrf": "^3.0.6",
