+++ pages/pil/training/content/species.js
@@ -1,0 +1,14 @@
+module.exports = {
+  title: 'What types of animals does this certificate cover?',
+  fields: {
+    species: {
+      label: '',
+      hint: 'Select all that apply.'
+    }
+  },
+  errors: {
+    species: {
+      required: 'Select one'
+    }
+  }
+};
