+++ pages/pil/training/routers/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   modules: require('./modules'),
   certificate: require('./certificate'),
-  exempt: require('./exempt')
+  exempt: require('./exempt'),
+  species: require('./species')
 };
