+++ pages/pil/training/routers/modules.js
@@ -1,10 +1,7 @@
 const { Router } = require('express');
 const form = require('../../../common/routers/form');
-const { processSpecies } = require('../../helpers');
 const { modules: schema } = require('../schema');
-const { pick, castArray } = require('lodash');
-const { buildModel, normalise } = require('../../../../lib/utils');
-const { modulesThatRequireSpecies } = require('../../constants');
+const { buildModel } = require('../../../../lib/utils');
 
 module.exports = settings => {
   const app = Router();
@@ -16,68 +13,10 @@ module.exports = settings => {
     next();
   });
 
-  app.use('/', form({
-    schema: {
-      ...schema,
-      ...schema.modules.options.reduce((obj, val) => {
-        return {
-          ...obj,
-          [`module-${normalise(val.value)}-species`]: val.species
+  app.use('/', form({ schema }));
 
-        };
-      }, {})
-    },
-    process: (req, res, next) => {
-      Object.assign(
-        (req.form.values = {
-          ...req.form.values,
-          ...processSpecies(req)
-        })
-      );
-      next();
-    },
-    locals: (req, res, next) => {
-      res.locals.static.schema = schema;
-      res.locals.static.modulesThatRequireSpecies = modulesThatRequireSpecies;
-      next();
-    }
-  }));
-
-  app.post('/', (req, res, next) => {
-    const fields = ['certificateNumber', 'accreditingBody', 'otherAccreditingBody', 'passDate', 'modules'];
-    const values = pick(req.session.form[req.model.id].values, fields);
-
-    values.modules = values.modules.map(module => {
-      if (!modulesThatRequireSpecies.includes(module)) {
-        return { module };
-      }
-
-      const species = req.form.values[`module-${normalise(module)}-species`];
-
-      if (!species) {
-        throw new Error(`Please select at least one type of animal for module ${module}`);
-      }
-
-      return { module, species: castArray(species).filter(s => s !== '') };
-    });
-
-    const opts = {
-      method: 'POST',
-      json: {
-        data: values
-      }
-    };
-
-    return req.api(`/establishment/${req.establishmentId}/profiles/${req.profileId}/certificate`, opts)
-      .then(() => {
-        delete req.session.form[req.model.id];
-        return next();
-      })
-      .catch(next);
-  });
-
-  app.post('/', (req, res, next) => {
-    return res.redirect(req.buildRoute('pil.update'));
+  app.post('/', (req, res) => {
+    res.redirect(req.buildRoute('pil.training.species'));
   });
 
   return app;
