+++ pages/pil/training/schema/index.js
@@ -1,5 +1,6 @@
 module.exports = {
   certificate: require('./certificate'),
-  modules: require('../../modules/schema'),
+  modules: require('./modules'),
+  species: require('./species'),
   exempt: require('../../exemptions/schema/exempt')
 };
