+++ pages/pil/training/schema/species.js
@@ -1,0 +1,12 @@
+const { toArray } = require('../../../../lib/utils');
+
+module.exports = {
+  species: {
+    inputType: 'speciesSelector',
+    format: toArray,
+    nullValue: [],
+    validate: [
+      'required'
+    ]
+  }
+};
